a big commit!
- some minor changes related to no-cache difficulties (does anybody parse out the fieldnames argument to no-cache??)

- move all routing logic out of the resources and into klein.  each resource is only responsible for determining what HTTP methods it allows.  hopefully this aides in testing.

NOTE: it's necessary to use a very, very recent klein as of 2015-11-24.  the master branch off github works great with python 3, but 15.1.0 (which is on PyPI) does not.  don't be fooled by the version in master!

TODO: there should be a distinction between immutable resources - onces that don't mutate instance state once created - and mutable resources.  klein lets us instantiate new ones for every request so this could be very handy.